requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# AI and OpenAI
openai==1.3.7

# Email
fastapi-mail==1.4.1

# File Storage and Media
# Google Cloud Storage
google-cloud-storage==2.10.0

# Google Services
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Calendar Integration
caldav==1.3.9
icalendar==5.0.11

# QR Code Generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Cryptography for Biometric Authentication
cryptography==41.0.7

# SMS Services
twilio==8.10.3

# Push Notifications
firebase-admin==6.4.0

# Rate Limiting and Circuit Breakers
slowapi==0.1.9
pybreaker==1.0.2
redis==5.0.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0